requirements.txt
streamlit 
requests 
pandas 
beautifulsoup4 
python-dotenv 
tenacity 
urllib3 
certifi
xlsxwriter